+++ examples/intsum/SingleLazy.hs
@@ -7,8 +7,8 @@ module SingleLazy (
 import Data.Int
 import qualified Data.Vector.Unboxed as U
 
-sumFrom1To :: Int32 -> Int32
+sumFrom1To :: Int64 -> Int64
 sumFrom1To mx = U.foldl' (+) 0 v
   where
-    v :: U.Vector Int32
+    v :: U.Vector Int64
     v = U.enumFromTo 1 mx
